feat(leif): move repository code to separate file
